readd call1 with new call_indirect1

+++ src/interp/interp.h
@@ -107,8 +107,8 @@ union web49_interp_opcode_t {
 struct web49_interp_block_t {
     web49_interp_opcode_t *code;
     uint32_t nlocals : 32;
-    uint16_t nparams : 16;
-    uint16_t nreturns : 15;
+    uint32_t nparams : 16;
+    uint32_t nreturns : 15;
     bool is_code : 1;
     union {
         struct {
